+++ src/include/Image.h
@@ -15,7 +15,7 @@ class RGBColor {
    RGBColor() = default;
    RGBColor(uint8_t red, uint8_t green, uint8_t blue)
        : r(red), g(green), b(blue) {}
-   RGBColor operator*(float scale);
+   RGBColor operator*(double scale);
    RGBColor operator+(int value);
 };
