+++ src/include/Mesh.h
@@ -7,17 +7,17 @@
 #include <string>
 #include <vector>
 
-using Vector3f = Eigen::Vector3f;
-using Vector4f = Eigen::Vector4f;
-using Vector2f = Eigen::Vector2f;
+using Vector3d = Eigen::Vector3d;
+using Vector4d = Eigen::Vector4d;
+using Vector2d = Eigen::Vector2d;
 using Vector3i = Eigen::Vector3i;
-using Matrix4f = Eigen::Matrix4f;
+using Matrix4d = Eigen::Matrix4d;
 
 class Mesh {
   protected:
-   std::vector<Vector3f> vertices;
-   std::vector<Vector3f> normals;
-   std::vector<Vector2f> uvs;
+   std::vector<Vector3d> vertices;
+   std::vector<Vector3d> normals;
+   std::vector<Vector2d> uvs;
    std::vector<std::vector<Vector3i>>
        faces;  // this Vec3i means vertex/uv/normal index
   public:
@@ -25,10 +25,10 @@ class Mesh {
    Mesh(const std::string& file_path);
    size_t vertices_nr();
    size_t faces_nr();
-   Vector3f normal(size_t iface, size_t nth_vertex);
-   Vector3f vertex(size_t ivertex);
-   Vector3f vertex(size_t iface, size_t nth_vertex);
-   void apply(Matrix4f transform);
+   Vector3d normal(size_t iface, size_t nth_vertex);
+   Vector3d vertex(size_t ivertex);
+   Vector3d vertex(size_t iface, size_t nth_vertex);
+   void apply(Matrix4d transform);
 };
 
 class Quad : public Mesh {
